Extract npm script for runSimulationsInParallelOn5Workers
