+++ README.md
@@ -25,6 +25,7 @@ npm run logXpLevelInfo
 npm run logXpTable
 npm run getLevelFor10kXP
 npm run oneSimulation
+npm run runSimulationsInParallelOn5Workers
 ```
 
 ## Formatting
