+++ package.json
@@ -12,7 +12,8 @@
     "logXpLevelInfo": "ts-node src/scripts/logXpLevelInfo.ts",
     "logXpTable": "ts-node src/scripts/logXpTable.ts",
     "getLevelFor10kXp": "ts-node src/scripts/getLevelFor10kXp.ts",
-    "oneSimulation": "ts-node src/scripts/oneSimulation.ts"
+    "oneSimulation": "ts-node src/scripts/oneSimulation.ts",
+    "runSimulationsInParallelOn5Workers": "tsc && node dist/scripts/runSimulationsInParallelOn5Workers.js"
   },
   "author": "Kyle Tolle",
   "license": "ISC",
